update all dependencies to latest versions
ureq 2 → 3: .set() → .header(), resp.into_reader() → resp.body_mut().as_reader(),
  resp.into_json() → resp.body_mut().read_json()
rusqlite 0.35 → 0.39: u64 no longer impl ToSql/FromSql, use i64 for timestamps
rustyline 15 → 17
sha2 0.10 → 0.11
rand 0.9 → 0.10: Rng::fill → RngExt::fill

+++ crates/tau/src/db.rs
@@ -17,7 +17,7 @@ pub struct StoredSession {
     pub model: Model,
     pub system_prompt: Option<String>,
     pub is_subscription: bool,
-    pub created_at: u64,
+    pub created_at: i64,
 }
 
 pub struct Db {
@@ -233,7 +233,7 @@ impl Db {
     pub fn append_message(&self, session_id: &str, message: &Message) -> crate::Result<()> {
         let json =
             serde_json::to_string(message).map_err(|e| crate::Error::Parse(e.to_string()))?;
-        let now = crate::types::timestamp_ms();
+        let now = crate::types::timestamp_ms() as i64;
         self.conn
             .execute(
                 "INSERT INTO messages (session_id, message_json, created_at) VALUES (?1, ?2, ?3)",
@@ -394,7 +394,7 @@ mod tests {
     #[test]
     fn list_sessions_ordered() {
         let db = Db::open_memory().unwrap();
-        for (id, ts) in [("s2", 2000u64), ("s1", 1000), ("s3", 3000)] {
+        for (id, ts) in [("s2", 2000i64), ("s1", 1000), ("s3", 3000)] {
             db.create_session(&StoredSession {
                 id: id.into(),
                 model: test_model(),
